tests/wpt/tests/css/css-values/calc-size/interpolate-size-computed.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>The interpolate-size property: computed values</title>
<link rel="help" href="https://drafts.csswg.org/css-values-5/#interpolate-size">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target"></div>
<script>
test_computed_value('interpolate-size', 'numeric-only');
test_computed_value('interpolate-size', 'allow-keywords');
</script>
